fix fks

+++ website/mysite/jubi/views.py
@@ -12,66 +12,64 @@ from datetime import date
 # Create your views here.
 @login_required(login_url='loginPage')
 def home(request):
-    search = request.POST.get('search')
     # Search for title if input is not empty
     if request.method == "POST":
+        # Check for new search
+        search = request.POST.get('search')
         if search:
-            title_set = Streaming.objects.filter(Q(title__icontains=search))
-            return render(request, 'search.html', {'title_set': title_set, 'search':search})
+            return render(request, 'search.html', {'tset': Streaming.objects.filter(title__name__icontains=search), 'search':search})
     return render(request, 'home.html', {})
 
 @login_required(login_url='loginPage')
 def title(request):
-    title = request.POST.get('title')
-    # Search for title if input is not empty
     if request.method == "POST":
+        # Check for new search
+        search = request.POST.get('search')
+        if search:
+            return render(request, 'search.html', {'tset': Streaming.objects.filter(title__name__icontains=search), 'search':search})
+
+        # Return title information
+        title = request.POST.get('title')
         if title:
-            title_set = Titles.objects.filter(Q(id=title))
-            streaming_set = Streaming.objects.filter(Q(title_id=title))
-            return render(request, 'title.html', {'title_set': title_set, 'streaming_set': streaming_set})
+            title_set = Streaming.objects.filter(title_id=title)
+            return render(request, 'title.html', {'tset': title_set, 'title': title_set[0]})
 
 @login_required(login_url='loginPage')
 def search(request):
-    search = request.POST.get('search')
     # Search for title if input is not empty
     if request.method == "POST":
+        search = request.POST.get('search')
+        wl_id = request.POST.get('add_wl')
+        w_id = request.POST.get('add_w')
+
+        # Check for new search
         if search:
-            title_set = Streaming.objects.filter(Q(name__icontains=search))
-            return render(request, 'search.html', {'title_set': title_set, 'search':search})
-        # Insert into watch later table
-        wl_id = request.POST.get('add_wl')
-        w_id = request.POST.get('add_w')
-        if wl_id:
-            add_wl = WatchLater(title_id=wl_id, user_id=request.user.id, date_added=date.today())
-            add_wl.save()
-        if w_id:
-            add_w = Watched(title_id = w_id, user_id=request.user.id)
-            add_w.save()
-    return render(request, 'search.html', {})
+            return render(request, 'search.html', {'tset': Streaming.objects.filter(title__name__icontains=search), 'search':search})
 
-
-# TODO: Fix so the page isn't reloaded every time
-def update_watched(request):
-    if request.method == "POST":
         # Insert into watch later table
-        wl_id = request.POST.get('add_wl')
-        w_id = request.POST.get('add_w')
         if wl_id:
-            add_wl = WatchLater(title_id=wl_id, user_id=request.user.id, date_added=date.today())
-            add_wl.save()
+            WatchLater(title_id=wl_id, user_id=request.user.id, date_added=date.today()).save()
+
+        # Insert into watched table
         if w_id:
-            add_w = Watched(title_id = w_id, user_id=request.user.id)
-            add_w.save()
-        return render(request, 'home.html', {})
+            Watched(title_id = w_id, user_id=request.user.id).save()
+
+    return render(request, 'home.html', {})
 
 @login_required(login_url='loginPage')
 def watch_later(request):
     # title_set = Titles.objects.filter(id=watch_later_set)
     if request.method == "POST":
+        # Check for new search
+        search = request.POST.get('search')
+        if search:
+            return render(request, 'search.html', {'tset': Streaming.objects.filter(title__name__icontains=search), 'search':search})
+
         # Delete from watch later table
         del_id = request.POST.get('btn_del')
         if del_id:
             WatchLater.objects.get(title_id=del_id, user_id=request.user.id).delete()
+        
         # Update priority of watch later record
         new_pri = request.POST.get('dp_pri')
         wl_id = request.POST.get('wl_id')
@@ -80,11 +78,31 @@ def watch_later(request):
             wl_obj.priority = new_pri
             wl_obj.save()
             
-    # TODO: Fix this later
-    watch_later_set = WatchLater.objects.filter(user_id=request.user.id)
-    title_set = Titles.objects.filter(id__in=watch_later_set.values_list("title_id"))
-    comb_set = zip(watch_later_set, title_set)
-    return render(request, 'watch_later.html', {'watch_later_set': watch_later_set,'comb_set': comb_set})
+    return render(request, 'watch_later.html', {'tset': WatchLater.objects.filter(user_id=request.user.id)})
+
+@login_required(login_url='loginPage')
+def watched(request):
+    # title_set = Titles.objects.filter(id=watch_later_set)
+    if request.method == "POST":
+        # Check for new search
+        search = request.POST.get('search')
+        if search:
+            return render(request, 'search.html', {'tset': Streaming.objects.filter(title__name__icontains=search), 'search':search})
+            
+        # Delete from watch later table
+        del_id = request.POST.get('btn_del')
+        if del_id:
+            Watched.objects.get(title_id=del_id, user_id=request.user.id).delete()
+
+        # Update priority of watch later record
+        new_rat = request.POST.get('dp_rat')
+        w_id = request.POST.get('w_id')
+        if new_rat:
+            wl_obj = Watched.objects.get(title_id=w_id, user_id=request.user.id)
+            wl_obj.rating = new_rat
+            wl_obj.save()
+
+    return render(request, 'watched.html', {'tset': Watched.objects.filter(user_id=request.user.id)})
 
 def registerPage(request):
     if request.user.is_authenticated:
@@ -121,24 +139,3 @@ def loginPage(request):
 def logoutUser(request):
     logout(request)
     return redirect('loginPage')
-
-@login_required(login_url='loginPage')
-def watched(request):
-    # title_set = Titles.objects.filter(id=watch_later_set)
-    if request.method == "POST":
-        # Delete from watch later table
-        del_id = request.POST.get('btn_del')
-        if del_id:
-            Watched.objects.get(title_id=del_id, user_id=request.user.id).delete()
-        # Update priority of watch later record
-        new_rat = request.POST.get('dp_rat')
-        w_id = request.POST.get('w_id')
-        if new_rat:
-            wl_obj = Watched.objects.get(title_id=w_id, user_id=request.user.id)
-            wl_obj.rating = new_rat
-            wl_obj.save()
-
-    watched_set = Watched.objects.filter(user_id=request.user.id)
-    title_set = Titles.objects.filter(id__in=watched_set.values_list("title_id"))
-    comb_set = zip(watched_set, title_set)
-    return render(request, 'watched.html', {'watched_set': watched_set,'comb_set': comb_set})
